+++ src/rpc_impl/rpc_disconnect_handlers.cc
@@ -25,7 +25,7 @@ void Rpc<Transport_>::handle_session_disconnect_req(SessionMgmtPkt *sm_pkt) {
   sprintf(issue_msg, "eRPC Rpc %s: Received disconnect request from %s. Issue",
           get_name().c_str(), sm_pkt->client.name().c_str());
 
-  uint32_t session_num = sm_pkt->server.session_num;
+  uint16_t session_num = sm_pkt->server.session_num;
   assert(session_num < session_vec.size());
 
   /* Check if the session was already disconnected */
@@ -66,7 +66,7 @@ void Rpc<Transport_>::handle_session_disconnect_resp(SessionMgmtPkt *sm_pkt) {
           sm_pkt->client.session_num);
 
   /* Try to locate the requester session for this response */
-  uint32_t session_num = sm_pkt->client.session_num;
+  uint16_t session_num = sm_pkt->client.session_num;
   assert(session_num < session_vec.size());
 
   Session *session = session_vec[session_num];
